initial version of releases.json (#2413)
first version of releases.json
