+++ packages/utils/src/utils.ts
@@ -1,4 +1,7 @@
 import { performance } from 'perf_hooks'
+import { validate } from 'jsonschema'
+import Hjson from 'hjson'
+import fs from 'fs'
 
 export function timer(fn: () => void): number {
   const start = performance.now()
@@ -20,4 +23,32 @@ export function expandVars(input: string, vars: { [key: string]: any }) {
     }
     return vars[varName]
   })
+}
+
+/**
+ * loads JSON data from file
+ * @param file_path json file to load
+ * @returns object parsed from JSON data
+ * @throws if unable to open the file the JSON data is malformed
+ */
+export function loadJson(file_path: string): any {
+  const content = fs.readFileSync(file_path, 'utf-8')
+  return Hjson.parse(content)
+}
+
+/**
+ * validates JSON data against JSON schema
+ * prints errors to the console and throws in case of non-conforming
+ * @param data parsed JSON data
+ * @param schema parsed JSON schema for the data
+ * @throws
+ */
+export function validateData(data: any, schema: any) {
+  const res = validate(data, schema)
+  for (const err of res.errors) {
+    console.log(err.stack)
+  }
+  if (res.errors.length > 0) {
+    throw new Error(`validation failed`)
+  }
 }
